backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
anthropic==0.52.0
cryptography==44.0.3
python-dotenv==1.1.0
pdfplumber>=0.11.0
python-magic-bin>=0.4.14
